Implemented Official Uno Scoring
Well, actually, it's not quite official yet. Instead of giving each player the value of their cards, it should give the winner the value of everyone else's cards.

I might just change the documentation to say I made this change to the rules. There's no point adding everyone else's if it's not a race to 500!

I might change it to say "Card Value Scoring", instead of "Official Uno Scoring"

I also:
• Changed the game options view a little bit
• Fixed the problem where the user could type whatever they wanted into the startup window player-type combo boxes
• Changed GameOptions to use auto-generated properties instead of fields (with validation on scoring system / stop after first)